[ur-6ash4] Add max_implement_cycles to ProjectConfig with server-wide fallback
Adds per-project max_implement_cycles to RawProjectConfig and ProjectConfig,
resolved as project override → server default → built-in default (6). Also
changes DEFAULT_MAX_IMPLEMENT_CYCLES and ServerConfig.max_implement_cycles
from i32 to u32 for consistency with max_fix_attempts, and updates all
construction sites and comparison casts accordingly.

+++ crates/server/src/workflow/handlers/dispatch_implement.rs
@@ -192,7 +192,7 @@ async fn check_cycle_limit(ctx: &WorkflowContext, ticket_id: &str) -> anyhow::Re
         .await?
         .ok_or_else(|| anyhow::anyhow!("no workflow found for ticket {ticket_id}"))?;
 
-    if workflow.implement_cycles >= max_cycles {
+    if workflow.implement_cycles as u32 >= max_cycles {
         let reason = format!(
             "implement cycle limit reached ({}/{})",
             workflow.implement_cycles, max_cycles
